fixed open window directory
