+++ ProjectJava/src/main/java/org/example/model/boardType/Board.java
@@ -1,0 +1,30 @@
+package org.example.model.boardType;
+
+import org.example.model.squares.SquareInfo;
+
+import java.util.ArrayList;
+
+public class Board {
+	private final ArrayList<SquareInfo> squares = new ArrayList<>();
+	private BoardType boardType;
+
+	public ArrayList<SquareInfo> getSquares() {
+		return squares;
+	}
+
+	public void addSquare(SquareInfo squareToAdd) {
+		squares.add(squareToAdd);
+	}
+
+	public BoardType getBoardType() {
+		return boardType;
+	}
+
+	public void setBoardType(BoardType boardType) {
+		this.boardType = boardType;
+	}
+
+	public boolean isWinner() {
+		return boardType.isWinner();
+	}
+}
