+++ ProjectJava/src/main/java/org/example/model/boardType/Bounce.java
@@ -5,8 +5,14 @@ import org.example.model.ui.Screen;
 
 public class Bounce implements BoardType {
 	boolean winner;
+	int lastSquare;
 
-	public void endAction(int posAfterRoll, int lastSquare, Player player) {
+	public Bounce(int squares) {
+		this.lastSquare = squares;
+	}
+
+	public void endAction(int posAfterRoll, Player player) {
+
 		if (posAfterRoll > lastSquare) {
 			posAfterRoll = 2 * lastSquare - posAfterRoll;
 			Screen.displayMessage("Hit Board and Bounced back " + (lastSquare - posAfterRoll) + " squares , new position is " + (posAfterRoll + 1));
